+++ mobile/android/android-components/components/support/utils/src/main/java/mozilla/components/support/utils/DownloadUtils.kt
@@ -126,6 +126,16 @@ object DownloadUtils {
     private val encodedSymbolPattern = Pattern.compile("%[0-9a-f]{2}|[0-9a-z!#$&+-.^_`|~]", Pattern.CASE_INSENSITIVE)
 
     /**
+     * Keep aligned with desktop generic content types:
+     * https://searchfox.org/mozilla-central/source/browser/components/downloads/DownloadsCommon.jsm#208
+     */
+    private val GENERIC_CONTENT_TYPES = arrayOf(
+        "application/octet-stream",
+        "binary/octet-stream",
+        "application/unknown"
+    )
+
+    /**
      * Guess the name of the file that should be downloaded.
      *
      * This method is largely identical to [android.webkit.URLUtil.guessFileName]
@@ -146,7 +156,11 @@ object DownloadUtils {
         val sanitizedMimeType = sanitizeMimeType(mimeType)
 
         val fileName = if (extractedFileName.contains('.')) {
-            changeExtension(extractedFileName, sanitizedMimeType)
+            if (GENERIC_CONTENT_TYPES.contains(mimeType)) {
+                extractedFileName
+            } else {
+                changeExtension(extractedFileName, sanitizedMimeType)
+            }
         } else {
             extractedFileName + createExtension(sanitizedMimeType)
         }